+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - will be 2.2.3]
+## [Unreleased - will be 2.3.0]
+
+### Added
+- Supports for new application passwords events (WordPress 5.6 and higher).
 
 ### Changed
 - Console now starts automatically in the admin dashboard if clicked.